Extract a Formatter::ConsoleCounts and use it everywhere
