+++ lib/cucumber/multiline_argument/data_table.rb
@@ -469,7 +469,8 @@ module Cucumber
 
         c = Cucumber::Term::ANSIColor.coloring?
         Cucumber::Term::ANSIColor.coloring = options[:color]
-        formatter = Formatter::Pretty.new(nil, io, options)
+        runtime = Struct.new(:configuration).new(Configuration.new)
+        formatter = Formatter::Pretty.new(runtime, io, options)
         formatter.instance_variable_set('@indent', options[:indent])
         Formatter::LegacyApi::Ast::MultilineArg.for(self).accept(Formatter::Fanout.new([formatter]))
         Cucumber::Term::ANSIColor.coloring = c
